% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_islh_ordinal}
\alias{scale_colour_islh_ordinal}
\alias{scale_color_islh_ordinal}
\title{Island Health ordinal colour scale}
\usage{
scale_colour_islh_ordinal(
  family = "blue",
  reverse = FALSE,
  ...,
  na.value = .islh_unknown()
)

scale_color_islh_ordinal(
  family = "blue",
  reverse = FALSE,
  ...,
  na.value = .islh_unknown()
)
}
\arguments{
\item{family}{Island Health colour family.}

\item{reverse}{Reverse the light-to-dark order.}

\item{...}{Additional arguments passed to \code{ggplot2::discrete_scale()}.}

\item{na.value}{Colour for missing values.}
}
\value{
A ggplot2 discrete scale.
}
\description{
Island Health ordinal colour scale
}
