% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_islh_signal}
\alias{scale_colour_islh_signal}
\alias{scale_color_islh_signal}
\title{Island Health signal colour scale}
\usage{
scale_colour_islh_signal(..., na.value = .islh_unknown())

scale_color_islh_signal(..., na.value = .islh_unknown())
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::scale_colour_manual()}.}

\item{na.value}{Colour for missing values.}
}
\value{
A ggplot2 discrete scale.
}
\description{
Island Health signal colour scale
}
