% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{islh_brand}
\alias{islh_brand}
\title{Look up a named Island Health brand colour}
\usage{
islh_brand(name)
}
\arguments{
\item{name}{One or more names from \code{.islh_brand_colours}.}
}
\value{
A character vector of hexadecimal colour values.
}
\description{
Look up a named Island Health brand colour
}
