% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{islh_example_plot}
\alias{islh_example_plot}
\title{Create a runnable Island Health example plot}
\usage{
islh_example_plot()
}
\value{
A ggplot object.
}
\description{
Create a runnable Island Health example plot
}
