% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-flextable.R
\name{islh_flextable}
\alias{islh_flextable}
\title{Apply Island Health styling to a flextable}
\usage{
islh_flextable(x, caption = NULL, autofit = TRUE, width = 1)
}
\arguments{
\item{x}{A flextable or data frame.}

\item{caption}{Optional caption.}

\item{autofit}{Automatically size columns to their contents.}

\item{width}{Table width as a fraction of the available text width.}
}
\value{
A styled flextable.
}
\description{
Apply Island Health styling to a flextable
}
