% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary.R
\name{islh_gtsummary_flex}
\alias{islh_gtsummary_flex}
\title{Convert a gtsummary table to a styled flextable}
\usage{
islh_gtsummary_flex(x, ...)
}
\arguments{
\item{x}{A gtsummary object.}

\item{...}{Additional arguments passed to \code{gtsummary::as_flex_table()}.}
}
\value{
A styled flextable.
}
\description{
Convert a gtsummary table to a styled flextable
}
