% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary.R
\name{islh_gtsummary_gt}
\alias{islh_gtsummary_gt}
\title{Convert a gtsummary table to a styled gt table}
\usage{
islh_gtsummary_gt(x, ..., embed_fonts = getOption("islh.embed_fonts", TRUE))
}
\arguments{
\item{x}{A gtsummary object.}

\item{...}{Additional arguments passed to \code{gtsummary::as_gt()}.}

\item{embed_fonts}{Embed BC Sans in the completed gt table. The session
default is set by \code{islh_setup()}.}
}
\value{
A styled gt table.
}
\description{
Convert a gtsummary table to a styled gt table
}
