% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{islh_hex}
\alias{islh_hex}
\title{Look up an Island Health brand colour}
\usage{
islh_hex(family, value)
}
\arguments{
\item{family}{One of the eight Island Health colour families.}

\item{value}{Brand lightness value.}
}
\value{
A character vector of hexadecimal colour values.
}
\description{
Look up an Island Health brand colour
}
