% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islhr-package.R
\docType{package}
\name{islhr-package}
\alias{islhr}
\alias{islhr-package}
\title{islhr: Island Health Branding for R Figures, Tables and Reports}
\description{
Applies Island Health brand colours, typography and layout to ggplot2 figures, flextable and gt tables, and Quarto HTML and Word reports. Includes project scaffolding and disclosure-control helpers used by the PHASE team.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/temuulene/islhr}
  \item Report bugs at \url{https://github.com/temuulene/islhr/issues}
}

}
\author{
\strong{Maintainer}: Temuulen Enebish \email{temuulen@gmail.com}

}
\keyword{internal}
