% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_islh}
\alias{scale_colour_islh}
\alias{scale_color_islh}
\title{Island Health qualitative colour scale}
\usage{
scale_colour_islh(reverse = FALSE, ..., na.value = .islh_unknown())

scale_color_islh(reverse = FALSE, ..., na.value = .islh_unknown())
}
\arguments{
\item{reverse}{Reverse the palette order.}

\item{...}{Additional arguments passed to \code{ggplot2::discrete_scale()}.}

\item{na.value}{Colour for missing values.}
}
\value{
A ggplot2 discrete scale.
}
\description{
Island Health qualitative colour scale
}
\examples{
\dontshow{assign("font", "", envir = getFromNamespace(".islh_state", "islhr"))}
ggplot2::ggplot(
  datasets::mtcars,
  ggplot2::aes(wt, mpg, colour = factor(cyl))
) +
  ggplot2::geom_point(size = 3) +
  scale_colour_islh() +
  ggplot2::labs(colour = "Cylinders") +
  theme_islh()
}
