% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_islh}
\alias{scale_fill_islh}
\title{Island Health qualitative fill scale}
\usage{
scale_fill_islh(reverse = FALSE, ..., na.value = .islh_unknown())
}
\arguments{
\item{reverse}{Reverse the palette order.}

\item{...}{Additional arguments passed to \code{ggplot2::discrete_scale()}.}

\item{na.value}{Colour for missing values.}
}
\value{
A ggplot2 discrete scale.
}
\description{
Island Health qualitative fill scale
}
