% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_islh_signal}
\alias{scale_fill_islh_signal}
\title{Island Health signal fill scale}
\usage{
scale_fill_islh_signal(..., na.value = .islh_unknown())
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::scale_fill_manual()}.}

\item{na.value}{Colour for missing values.}
}
\value{
A ggplot2 discrete scale.
}
\description{
Island Health signal fill scale
}
