% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_islh}
\alias{theme_islh}
\title{Island Health ggplot theme}
\usage{
theme_islh(base_size = 12, grid = c("y", "x", "both", "none"))
}
\arguments{
\item{base_size}{Base font size in points.}

\item{grid}{Major gridlines to display: y, x, both, or none.}
}
\value{
A ggplot2 theme object.
}
\description{
Island Health ggplot theme
}
